+++ app/views/signup/profile.html.erb
@@ -68,7 +68,7 @@
 
     <div class="checkbox">
       <label>
-        <%= f.check_box :newsletter %>
+        <%= f.check_box :newsletter, checked: true %>
         <%= t :".keep_me_informed" %>
       </label>
     </div>
