+++ spec/handlers/signup_profile_instructor_spec.rb
@@ -37,6 +37,10 @@ RSpec.describe SignupProfileInstructor, type: :handler do
 
     end
 
+    it "requires a number >= 0 for num_students" do
+      skip # TODO
+    end
+
     context "when the fields are properly filled in" do
       let (:the_call) { -> {
         described_class.handle(
